+++ app/src/main/java/com/saarlabs/tminus/android/widget/MBTAStationBoardWidget.kt
@@ -53,6 +53,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlin.math.max
 import kotlin.math.pow
+import com.saarlabs.tminus.android.util.clockWithTrack
 
 public class MBTAStationBoardWidget : GlanceAppWidget() {